pyproject.toml
[project]
name = "bachata-choreography-generator"
version = "0.1.0"
description = "AI-powered Bachata choreography generator that creates dance sequences from YouTube music"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Django core
    "django>=5.2,<5.3",
    "django-htmx>=1.17.0",
    "gunicorn>=21.2.0",
    "whitenoise>=6.6.0",
    
    # Django REST Framework (for microservices API)
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "drf-spectacular>=0.27.0",
    "django-cors-headers>=4.3.0",
    
    # Database
    "psycopg2-binary>=2.9.0",
    
    # Video/Audio processing
    "yt-dlp>=2023.1.0",
    "librosa>=0.9.0",
    "opencv-python>=4.5.0",
    "ultralytics>=8.0.0",  # YOLOv8 for pose detection
    
    # ML/AI (runtime)
    "numpy>=1.21.0",
    "scipy>=1.11.0",
    "torch>=2.0.0",
    "sentence-transformers>=2.0.0",
    "google-generativeai>=0.3.0",
    
    # Search
    "elasticsearch>=8.11.0",
    
    # Cloud services
    "google-cloud-secret-manager>=2.0.0",
    "google-cloud-storage>=2.10.0",
    "google-cloud-run>=0.10.0",
    "django-storages[google]>=1.14.0",
    
    # Monitoring
    "elastic-apm>=6.19.0",
    
    # Utilities
    "python-dotenv>=1.1.1",
    "tqdm>=4.62.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-django>=4.5.0",
    "pytest-timeout>=2.1.0",
]
